refactor code structure for toolbar
